third_party/WebKit/LayoutTests/fast/dom/custom/document-register-svg-extends.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function () {
    var proto = Object.create(SVGGElement.prototype);

    assert_throws('NOT_SUPPORTED_ERR', function() {
        document.registerElement('x-foo', {prototype: proto});
    }, 'register of an SVG element requires the extends parameter');

    var XFoo = document.registerElement('x-foo', {prototype: proto, extends: 'g'});

    var created = new XFoo();
    assert_true(created instanceof SVGGElement, 'created element should be an SVG element');

}, 'registration of SVG types');
</script>
